refactored and added packages.
Organized classes into relative packages.
